+++ localization/pose_instability_detector/CMakeLists.txt
@@ -5,8 +5,8 @@ find_package(autoware_cmake REQUIRED)
 autoware_package()
 
 ament_auto_add_executable(pose_instability_detector
-  src/pose_instability_detector_node.cpp
-  src/pose_instability_detector_core.cpp
+  src/main.cpp
+  src/pose_instability_detector.cpp
 )
 
 ament_auto_package(
