+++ localization/pose_instability_detector/src/pose_instability_detector.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "pose_instability_detector_core.hpp"
+#include "pose_instability_detector.hpp"
 
 PoseInstabilityDetector::PoseInstabilityDetector() : Node("pose_instability_detector")
 {
